Revert "rename k8dash to skooner"
This reverts commit eaff4a6b9a4711c4e28849d7e77cd4c6cf2b57e8.
