+++ CONTRIBUTING.md
@@ -1,4 +1,4 @@
-# Contributing to Skooner
+# Contributing to K8dash
 We welcome any kinds of contribution and want to make contributing to this project as easy and transparent as
 possible.
