+++ client/public/index.html
@@ -17,7 +17,7 @@
     <link rel="manifest" href="./manifest.json" />
     <link rel="apple-touch-icon" sizes="512x512 192x192 144x144 64x64 32x32 24x24 16x16" href="./logo.png">
     <link rel="apple-touch-startup-image" href="./logo.png">
-    <title>Skooner</title>
+    <title>K8Dash</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
