+++ client/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "Skooner",
-  "name": "Skooner - Kubernetes Dashboard",
+  "short_name": "K8Dash",
+  "name": "K8Dash - Kubernetes Dashboard",
   "icons": [
     {
       "src": "logo.png",
