+++ client/src/art/k8dashSvg.tsx
